+++ projects/ngx-draggable-dom/src/lib/classes/ngx-draggable-bounds-check-event.ts
@@ -4,25 +4,25 @@
  */
 export class NgxDraggableBoundsCheckEvent {
 
-    public top: boolean;
-    public right: boolean;
-    public bottom: boolean;
-    public left: boolean;
+  public readonly top: boolean;
+  public readonly right: boolean;
+  public readonly bottom: boolean;
+  public readonly left: boolean;
 
-    /**
-     * Read only property that indicates if one or more of the boundaries have been collided with.
-     *
-     * @return True if any of the boundaries are collided.
-     */
-    public get hasCollision(): boolean {
-        return !!this.top || !!this.right || !!this.bottom || !!this.left;
-    }
+  /**
+   * Read only property that indicates if one or more of the boundaries have been collided with.
+   *
+   * @return True if any of the boundaries are collided.
+   */
+  public get hasCollision(): boolean {
+    return !!this.top || !!this.right || !!this.bottom || !!this.left;
+  }
 
-    constructor(top: boolean, right: boolean, bottom: boolean, left: boolean) {
-        this.top = (!!top) ? top : false;
-        this.right = (!!right) ? right : false;
-        this.bottom = (!!bottom) ? bottom : false;
-        this.left = (!!left) ? left : false;
-    }
+  constructor(top: boolean, right: boolean, bottom: boolean, left: boolean) {
+    this.top = (!!top) ? top : false;
+    this.right = (!!right) ? right : false;
+    this.bottom = (!!bottom) ? bottom : false;
+    this.left = (!!left) ? left : false;
+  }
 
 }
