+++ projects/ngx-draggable-dom/tslint.json
@@ -116,7 +116,10 @@
     "no-output-on-prefix": true,
     "no-input-rename": true,
     "no-output-rename": true,
+    "use-host-property-decorator": true,
+    "use-input-property-decorator": true,
     "use-life-cycle-interface": true,
+    "use-output-property-decorator": true,
     "use-pipe-transform-interface": true,
     "component-class-suffix": true,
     "directive-class-suffix": true
